+++ jannovar-cli/src/main/java/de/charite/compbio/jannovar/cmd/annotate_pos/AnnotatePositionCommand.java
@@ -43,7 +43,7 @@ public class AnnotatePositionCommand extends JannovarAnnotationCommand {
 	@Override
 	public void run() throws JannovarException {
 		System.err.println("Options");
-		options.print();
+		options.print(System.err);
 
 		System.err.println("Deserializing transcripts...");
 		deserializeTranscriptDefinitionFile();