+++ jannovar-cli/src/main/java/de/charite/compbio/jannovar/cmd/annotate_pos/AnnotatePositionCommandLineParser.java
@@ -24,6 +24,7 @@ public class AnnotatePositionCommandLineParser extends JannovarAnnotationCommand
 
 		// Fill the resulting JannovarOptions.
 		JannovarOptions result = new JannovarOptions();
+		result.printProgressBars = true;
 		result.command = JannovarOptions.Command.ANNOTATE_POSITION;
 
 		if (cmd.hasOption("help")) {