+++ jannovar-cli/src/main/java/de/charite/compbio/jannovar/cmd/annotate_vcf/AnnotateVCFCommandLineParser.java
@@ -23,6 +23,7 @@ public class AnnotateVCFCommandLineParser extends JannovarAnnotationCommandLineP
 
 		// Fill the resulting JannovarOptions.
 		JannovarOptions result = new JannovarOptions();
+		result.printProgressBars = true;
 		result.command = JannovarOptions.Command.ANNOTATE_VCF;
 
 		if (cmd.hasOption("help")) {