+++ jannovar-cli/src/main/java/de/charite/compbio/jannovar/cmd/db_list/DatabaseListCommandLineParser.java
@@ -57,6 +57,7 @@ public class DatabaseListCommandLineParser {
 
 		// Fill the resulting JannovarOptions.
 		JannovarOptions result = new JannovarOptions();
+		result.printProgressBars = true;
 		result.command = JannovarOptions.Command.DB_LIST;
 
 		if (cmd.hasOption("help")) {