+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 **develop**
 
+* Using log4j/slf4j for I/O in jannovar-core.
+* Adding `PrintStream` as parameter to `JannovarOptions#print`.
 * Compressing serialized file.
 * Fixing problems with block substitution (delins) case (#87).
 * Changing namespace to `de.charite.compbio.jannovar`.